+++ packages/loaders/code/__tests__/assets/bar.ts
@@ -1,0 +1,7 @@
+import gql from 'graphql-tag';
+
+export const typeDefs = gql`
+  type Mutation {
+    bar: String
+  }
+`;